requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
python-dotenv==1.0.0
streamlit==1.28.0
lxml==4.9.3